+++ src/main/java/com/appdynamics/extensions/aws/apigateway/schemas/TraditionalMetricSchema.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2018 AppDynamics,Inc.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.appdynamics.extensions.aws.apigateway.schemas;
+
+/**
+ * Created by venkata.konala on 5/27/18.
+ */
+public class TraditionalMetricSchema {
+
+    private String metricName;
+
+    private String metricPath;
+
+    private String metricValue;
+
+
+    public String getMetricName() {
+        return metricName;
+    }
+
+    public void setMetricName(String metricName) {
+        this.metricName = metricName;
+    }
+
+    public String getMetricPath() {
+        return metricPath;
+    }
+
+    public void setMetricPath(String metricPath) {
+        this.metricPath = metricPath;
+    }
+
+    public String getMetricValue() {
+        return metricValue;
+    }
+
+    public void setMetricValue(String metricValue) {
+        this.metricValue = metricValue;
+    }
+}