+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # AppDynamics API Gateway Monitoring Extension CHANGELOG
 
-## 1.1.0 - Apr 23, 2018
+## 2.0.0 - May 27, 2018
+1. Added capability to send standard cloudwatch metrics to AppDynamics eventsService.
+2. Added infrastructure metrics related to API Gateway, that can only be accessed via the AppDynamics eventService.
+
+## 1.0.0 - May 24, 2018
 1. Initial version.